신청 모델 추가, 리소스 변경에 따른 상태 관리, API 요청을 Register와 관련된 주체가 수행
