--- src/hooks/useMemberStore.js
@@ -1,6 +1,0 @@
-import { memberStore } from '../stores/MemberStore';
-import useStore from './useStore';
-
-export default function useMemberStore() {
-  return useStore(memberStore);
-}
