+++ src/stores/RegisterStore.js
@@ -1,3 +1,5 @@
+/* eslint-disable class-methods-use-this */
+
 import { registerApiService } from '../services/RegisterApiService';
 import Store from './Store';
 
@@ -7,6 +9,21 @@ export default class RegisterStore extends Store {
 
     this.registeredGameId = -1;
     this.registerErrorCodeAndMessage = {};
+
+    this.members = [];
+    this.membersErrorMessage = '';
+  }
+
+  async fetchMembers(gameId) {
+    try {
+      const data = await registerApiService.fetchMembers(gameId);
+      this.members = data.members;
+      this.publish();
+    } catch (error) {
+      const { errorMessage } = error.response.data;
+      this.membersErrorMessage = errorMessage;
+      this.publish();
+    }
   }
 
   async registerToGame(gameId) {
@@ -21,6 +38,10 @@ export default class RegisterStore extends Store {
       };
       return '';
     }
+  }
+
+  async cancelParticipateGame(gameId) {
+    await registerApiService.cancelParticipateGame(gameId);
   }
 }
